+++ includes/shortcodes/renderers/call_to_action.php
@@ -5,7 +5,7 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.4.1
+ * @version  1.4.2
  *
  * @uses  $codes_globals['sizes']['values']
  *
@@ -83,8 +83,10 @@
 
 // Output
 
-	if ( ! empty( $atts['caption'] . $atts['content'] ) ) {
-		$output = '<div class="' . esc_attr( $atts['class'] ) . '"><div class="wm-call-to-action-content wm-call-to-action-element">' . $atts['caption'] . $atts['content'] . '</div>' . $atts['button'] . '</div>';
+	$shortcode_output = $atts['caption'] . $atts['content'];
+
+	if ( ! empty( $shortcode_output ) ) {
+		$output = '<div class="' . esc_attr( $atts['class'] ) . '"><div class="wm-call-to-action-content wm-call-to-action-element">' . $shortcode_output . '</div>' . $atts['button'] . '</div>';
 	} else {
 		$output = esc_html__( 'Sorry, there is nothing to display here&hellip;', 'webman-amplifier' );
 	}